+++ trunk/Simulation/src/SystemGutsRep.h
@@ -94,7 +94,7 @@ public:
     const State& getDefaultState() const {return defaultState;}
     State&       updDefaultState()       {return defaultState;}
 
-    int              getNSubsystems()            const {return subsystems.size();}
+    int              getNSubsystems()               const {return (int)subsystems.size();}
     const Subsystem& getSubsystem(SubsystemIndex i) const {return subsystems[i];}
     Subsystem&       updSubsystem(SubsystemIndex i)       {return subsystems[i];}
